Методы для книг перенесены в компонент Book в папке API
