+++ react-web-client/src/redux/state.js
@@ -1,10 +1,10 @@
-import * as axios from "axios";
 import { getPainterStyle } from "./API/PainterStyles";
 import { getPainterByCount } from "./API/Painter";
 import { getPublishersByCount } from "./API/Publisher";
 import { getCoverTypes } from "./API/CoverTypes";
 import { getInterpretersByCount } from "./API/Interpreter";
 import { getAuthorByCount } from "./API/Author";
+import { getBooksByCount } from "./API/Book";
 
 let state = {
     contentPage: {
@@ -41,106 +41,6 @@ export let initializeData = () => {
     getBooksByCount(10);
 }
 
-// TODO: вынести в отдельный компонент
-export let getBooksByCount = (count) => {
-    axios.get('https://localhost:44394/api/Book/GetBooks/take/' + count + '/skip/0').then(response => {
-        console.log('get books', response);
-        if (state.contentPage.books.length < 1) {
-            state.contentPage.books = response.data.previewBooks;
-        } else {
-            response.data.previewBooks.forEach(element => {
-                state.contentPage.books.push(element);
-            });
-        }
-    })
-    return state.contentPage.books;
-}
-
-export let addBook = (name, publishDate, coverTypeName, description, format, countPage, price, weight, duplicate, ageLimit, publisherName) => {
-    let coverTypeId = state.contentPage.coverTypes.find(ct => ct.name === coverTypeName).id;
-    let publisherId = state.contentPage.publishers.find(p => p.name === publisherName).id;
-
-    let book = {
-        id: newGuid(),
-        name,
-        publishDate,
-        coverTypeId,
-        genreId: newGuid(),
-        languageId: newGuid(),
-        description,
-        isbN_13: '1',
-        format,
-        countPage,
-        price,
-        weight,
-        duplicate,
-        ageLimit,
-        publisherId,
-        authorsIds: [
-            newGuid()
-        ],
-        paintersIds: [
-            newGuid()
-        ],
-        interpretersIds: [
-            newGuid()
-        ]
-    }
-    axios.post('https://localhost:44394/api/Book', book).then(response => {
-        console.log('add book', response);
-    })
-
-    getBooksByCount(10);
-    // rerenderEntireTree(state);
-}
-
-export let updateBook = (oldName, name, publishDate, coverTypeName, description, format, countPage, price, weight, duplicate, ageLimit, publisherName) => {
-    let oldBook = state.contentPage.books.find(b => b.name === oldName)
-
-    let coverTypeId = state.contentPage.coverTypes.find(ct => ct.name === coverTypeName).id;
-    let publisherId = state.contentPage.publishers.find(p => p.name === publisherName).id;
-
-    let book = {
-        id: oldBook.id,
-        name,
-        publishDate,
-        coverTypeId,
-        genreId: newGuid(),
-        languageId: newGuid(),
-        description,
-        isbN_13: '1',
-        format,
-        countPage,
-        price,
-        weight,
-        duplicate,
-        ageLimit,
-        publisherId,
-        authorsIds: [
-            newGuid()
-        ],
-        paintersIds: [
-            newGuid()
-        ],
-        interpretersIds: [
-            newGuid()
-        ]
-    }
-    axios.put('https://localhost:44394/api/Book/', book).then(response => {
-        console.log('update book', response);
-    })
-
-    initializeData();
-}
-
-export let deleteBook = (name) => {
-    let bookId = state.contentPage.books.find(ps => ps.name === name).id
-
-    axios.delete('https://localhost:44394/api/Book/' + bookId).then(response => {
-        console.log('delete book', response);
-    })
-    initializeData();
-}
 
 export let newGuid = () => {
     return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
